requirements.txt
Django==1.6.1
South==0.8.4
MySQL-python==1.2.4
django_compressor==1.3
pytz>2012
Pillow==2.2.1
raven
ydcommon==0.1.7
django-tinymce==1.5.2
flup==1.0.3.dev-20110405
django-grappelli==2.5.0
django-filebrowser==3.5.2
django-celery==3.0.11
pyflakes
django-passwords==0.2.0
django-localflavor==1.0